Created empty builder generator
