+++ build-tools/processor/src/main/java/co/elastic/apm/compile/processor/generators/InstrumentationBuilderGenerator.java
@@ -1,0 +1,13 @@
+package co.elastic.apm.compile.processor.generators;
+
+import java.util.List;
+
+import javax.lang.model.element.TypeElement;
+
+public class InstrumentationBuilderGenerator implements Generator {
+
+    @Override
+    public Result generate(List<TypeElement> from) {
+        return null;
+    }
+}
